bakes

+++ database/migrations/2021_05_30_070247_create_cart_table.php
@@ -15,8 +15,8 @@ class CreateCartTable extends Migration
     {
         Schema::create('cart', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id');
-            $table->integer('product_id');
+            $table->bigInteger('user_id');
+            $table->bigInteger('product_id');
             $table->timestamps();
         });
     }
